Ran migrate, needed to chance the value it was checking for to integert

+++ data/migrations/01-make_cars_table.js
@@ -5,7 +5,7 @@ exports.up = function (knex) {
     table.string("vin", 17).notNullable().unique()
     table.string("make", 100).notNullable()
     table.string("model", 100).notNullable()
-    table.numeric("mileage", 999999).notNullable()
+    table.integer("mileage", 999999).notNullable()
     table.string("title", 100)
     table.string("transmission", 100)
   })
